Moved static logic from SSTable and LsmDAO (climate fixes)
