+++ src/main/java/ru/mail/polis/lsm/artem_drozdov/util/File.java
@@ -1,9 +1,11 @@
 package ru.mail.polis.lsm.artem_drozdov.util;
 
 import java.io.IOException;
+import java.nio.channels.FileChannel;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.nio.file.StandardOpenOption;
 
 public final class File {
     private File() {
@@ -25,5 +27,14 @@ public final class File {
 
     public static Path getTmpFile(Path file) {
         return resolveWithExt(file, ".tmp");
+    }
+
+    public static FileChannel openForWrite(Path tmpFileName) throws IOException {
+        return FileChannel.open(
+                tmpFileName,
+                StandardOpenOption.CREATE_NEW,
+                StandardOpenOption.WRITE,
+                StandardOpenOption.TRUNCATE_EXISTING
+        );
     }
 }
